requirements.txt
apt-clone==0.2.1
apturl==0.5.2
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
ccsm==0.9.14.1
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.0.1
colorama==0.4.3
command-not-found==0.3
compizconfig-python==0.9.14.1
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
dominate==2.6.0
entrypoints==0.3
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
grpcio==1.16.1
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
IMDbPY==6.8
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
netaddr==0.7.19
netifaces==0.10.4
nose==1.3.7
oauthlib==3.1.0
onboard==1.4.1
packaging==20.3
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-debian===0.1.36ubuntu1
python-magic==0.4.16
python-xapp==2.2.1
python-xlib==0.23
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.4.21
systemd-python==234
tldextract==2.2.1
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
visitor==0.1.3
wadllib==1.3.3
Werkzeug==2.0.1
WTForms==2.3.3
xkit==0.0.0
youtube-dl==2021.4.26
zeroconf==0.24.4

python-dotenv~=0.18.0